[ISSUE #28332] code review

+++ airbyte-cdk/python/airbyte_cdk/connector_builder/migrator/source.py
@@ -7,13 +7,9 @@ import logging
 import os
 import re
 from enum import Enum
-from typing import Any, Dict, List, Optional
+from typing import Any, Dict, List, Optional, Tuple
 
-import yaml
-import yaml.emitter
-import yaml.representer
-import yaml.resolver
-import yaml.serializer
+import ruamel.yaml
 from airbyte_cdk.sources.declarative.models.declarative_component_schema import AuthFlow, Spec
 
 logger = logging.getLogger("migrator.source")
@@ -28,8 +24,10 @@ class SourceRepository:
         spec = self._fetch_spec_outside_manifest(source_name)
         if spec:
             manifest_path = self._manifest_path(source_name)
+            yaml = ruamel.yaml.YAML()
+            yaml.representer.add_multi_representer(Enum, lambda dumper, enum: dumper.represent_data(enum.value))
             with open(os.path.join(manifest_path)) as f:
-                manifest = yaml.safe_load(f.read())
+                manifest = yaml.load(f.read())
             if "spec" in manifest:
                 # It should be fine to ignore since the CATs test `test_match_expected` ensure that both are the same
                 logger.info(
@@ -39,21 +37,20 @@ class SourceRepository:
                 manifest["spec"] = spec.dict(exclude_none=True)
 
             with open(manifest_path, 'w') as outfile:
-                yaml.dump(manifest, outfile, sort_keys=False, Dumper=_ManifestDumper)
+                yaml.dump(manifest, outfile)
             self._delete_spec_outside_manifest(source_name)
         else:
             logger.debug(f"Source {source_name} does not have a spec outside the manifest.yaml")
 
     def _fetch_spec_outside_manifest(self, source_name: str) -> Optional[Spec]:
-        json_spec_path = os.path.join(self._python_package_path(source_name), "spec.json")
-        yaml_spec_path = os.path.join(self._python_package_path(source_name), "spec.yaml")
+        json_spec_path, yaml_spec_path = self._get_spec_path(source_name)
 
-        if os.path.exists(json_spec_path):
+        if json_spec_path and os.path.exists(json_spec_path):
             with open(json_spec_path) as json_file:
                 return self._assemble_spec(json.load(json_file))
-        elif os.path.exists(yaml_spec_path):
+        elif yaml_spec_path and os.path.exists(yaml_spec_path):
             with open(yaml_spec_path) as yaml_file:
-                return self._assemble_spec(yaml.safe_load(yaml_file))
+                return self._assemble_spec(ruamel.yaml.load(yaml_file.read(), Loader=ruamel.yaml.RoundTripLoader))
         else:
             return None
 
@@ -67,12 +64,11 @@ class SourceRepository:
         )
 
     def _delete_spec_outside_manifest(self, source_name: str) -> None:
-        json_spec_path = os.path.join(self._python_package_path(source_name), "spec.json")
-        yaml_spec_path = os.path.join(self._python_package_path(source_name), "spec.yaml")
+        json_spec_path, yaml_spec_path = self._get_spec_path(source_name)
 
-        if os.path.exists(json_spec_path):
+        if json_spec_path and os.path.exists(json_spec_path):
             os.remove(json_spec_path)
-        elif os.path.exists(yaml_spec_path):
+        elif yaml_spec_path and os.path.exists(yaml_spec_path):
             os.remove(yaml_spec_path)
 
     def fetch_no_code_sources(self) -> List[str]:
@@ -96,6 +92,15 @@ class SourceRepository:
             # `class_name`, it'll be a custom component
             return not re.search("class_name:", file_content)
 
+    def _get_spec_path(self, source_name: str) -> Tuple[Optional[str], Optional[str]]:
+        json_spec_path = os.path.join(self._python_package_path(source_name), "spec.json")
+        yaml_spec_path = os.path.join(self._python_package_path(source_name), "spec.yaml")
+
+        return (
+            json_spec_path if os.path.exists(json_spec_path) else None,
+            yaml_spec_path if os.path.exists(yaml_spec_path) else None
+        )
+
     def _manifest_path(self, source_name: str) -> str:
         return os.path.join(self._python_package_path(source_name), "manifest.yaml")
 
@@ -104,63 +109,3 @@ class SourceRepository:
 
     def _path_from_name(self, source_name: str) -> str:
         return os.path.join(self._base_path, "airbyte-integrations", "connectors", source_name)
-
-
-class _IndentingEmitter(yaml.emitter.Emitter):
-    def increase_indent(self, flow: bool = False, indentless: bool = False) -> None:
-        """Ensure that lists items are always indented."""
-        super().increase_indent(flow=False, indentless=False)  # type: ignore
-
-
-class _ManifestDumper(
-    _IndentingEmitter,
-    yaml.serializer.Serializer,
-    yaml.representer.Representer,
-    yaml.resolver.Resolver,
-):
-    def __init__(  # type: ignore
-        self,
-        stream,
-        default_style=None,
-        default_flow_style=False,
-        canonical=None,
-        indent=None,
-        width=None,
-        allow_unicode=None,
-        line_break=None,
-        encoding=None,
-        explicit_start=None,
-        explicit_end=None,
-        version=None,
-        tags=None,
-        sort_keys=True,
-    ) -> None:
-        _IndentingEmitter.__init__(
-            self,
-            stream,
-            canonical=canonical,
-            indent=indent,
-            width=width,
-            allow_unicode=allow_unicode,
-            line_break=line_break,
-        )
-        yaml.serializer.Serializer.__init__(
-            self,
-            encoding=encoding,
-            explicit_start=explicit_start,
-            explicit_end=explicit_end,
-            version=version,
-            tags=tags,
-        )
-        yaml.representer.Representer.__init__(
-            self,
-            default_style=default_style,
-            default_flow_style=default_flow_style,
-            sort_keys=sort_keys,
-        )
-        yaml.resolver.Resolver.__init__(self)
-
-    def represent_data(self, data: Any) -> yaml.Node:
-        if isinstance(data, Enum):
-            return self.represent_data(data.value)
-        return super().represent_data(data)
